View, create and edit request
