--- application/Controller/SongsController.php
@@ -1,179 +1,0 @@
-<?php
-
-/**
- * Class SongsController
- * This is a demo Controller class.
- *
- * If you want, you can use multiple Models or Controllers.
- *
- * Please note:
- * Don't use the same name for class and method, as this might trigger an (unintended) __construct of the class.
- * This is really weird behaviour, but documented here: http://php.net/manual/en/language.oop5.decon.php
- *
- */
-
-namespace Mini\Controller;
-
-use Mini\Core\Controller;
-use Mini\Core\Redirect;
-use Mini\Core\Request;
-use Mini\Model\Song;
-
-class SongsController extends Controller
-{
-    /**
-     * Construct this object by extending the basic Controller class.
-     */
-    // public function __construct()
-    // {
-    //     parent::__construct();
-
-    //     // VERY IMPORTANT: All controllers/areas that should only be usable by logged-in users
-    //     // need this line! Otherwise not-logged in users could do actions.
-    //     Auth::checkAuthentication();
-    // }
-
-    /**
-     * PAGE: index
-     * This method handles what happens when you move to http://yourproject/songs/index
-     */
-    public function index()
-    {
-        // Instance new Model (Song)
-        $Song = new Song();
-        // getting all songs and amount of songs
-        $songs = $Song->getAllSongs();
-        $amount_of_songs = $Song->getAmountOfSongs();
-
-        // load views. within the views we can echo out $songs and $amount_of_songs easily
-        $this->View->render('songs/index', array('songs' => $songs, 'amount_of_songs' => $amount_of_songs));
-    }
-
-    /**
-     * ACTION: Show the form for creating a new resource.
-     *
-     * This method handles what happens when you move to http://yourproject/songs/addsong
-     * IMPORTANT: This is not a normal page, it's an ACTION. This is where the "add a song" form on songs/index
-     * directs the user after the form submit. This method handles all the POST data from the form and then redirects
-     * the user back to songs/index via the last line: header(...)
-     * This is an example of how to handle a POST request.
-     */
-    public function create()
-    {
-        // if we have POST data to create a new song entry
-        if (Request::isset("submit_add_song")) {
-            // Instance new Model (Song)
-            $Song = new Song();
-            // do addSong() in model/model.php
-            $Song->addSong(Request::post('artist'), Request::post('track'), Request::post('link'));
-        }
-
-        // where to go after song has been added
-        Redirect::to('songs/index');
-    }
-
-    /**
-     * Create a new song object and stores to the database.
-     *
-     * @return void
-     */
-    public function store() {
-
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id) {
-
-    }
-
-     /**
-     * ACTION: Show the form for editing the specified resource.
-     * This method handles what happens when you move to http://yourproject/songs/editsong
-     * @param int $id Id of the to-edit song
-     */
-    public function edit($id)
-    {
-        // if we have an id of a song that should be edited
-        if (isset($id)) {
-            // Instance new Model (Song)
-            $Song = new Song();
-            // do getSong() in model/model.php
-            $song = $Song->getSong($id);
-
-            // in a real application we would also check if this db entry exists and therefore show the result or
-            // redirect the user to an error page or similar
-
-            // load views. within the views we can echo out $song easily
-            $this->View->render('songs/edit', array('song' => $song));
-        } else {
-            // redirect user to songs index page (as we don't have a song_id)
-            Redirect::to('songs/index');
-        }
-    }
-
-    /**
-     * ACTION: Update the specified resource in storage.
-     *
-     * This method handles what happens when you move to http://yourproject/songs/updatesong
-     * IMPORTANT: This is not a normal page, it's an ACTION. This is where the "update a song" form on songs/edit
-     * directs the user after the form submit. This method handles all the POST data from the form and then redirects
-     * the user back to songs/index via the last line: header(...)
-     * This is an example of how to handle a POST request.
-     */
-    public function update($id)
-    {
-        // if we have POST data to create a new song entry
-        if (Request::isset("submit_update_song")) {
-            // Instance new Model (Song)
-            $Song = new Song();
-            // do updateSong() from model/model.php
-            $Song->updateSong(Request::post('artist'), Request::post('track'), Request::post('link'), Request::post('song_id'));
-        }
-
-        // where to go after song has been added
-        Redirect::to('songs/index');
-    }
-
-    /**
-     * AJAX-ACTION: ajaxGetStats
-     * TODO documentation
-     */
-    public function ajaxGetStats()
-    {
-        // Instance new Model (Song)
-        $Song = new Song();
-        $amount_of_songs = $Song->getAmountOfSongs();
-
-        // simply echo out something. A supersimple API would be possible by echoing JSON here
-        echo $amount_of_songs;
-    }
-
-    /**
-     * ACTION: Remove the specified resource from storage.
-     * 
-     * This method handles what happens when you move to http://yourproject/songs/deletesong
-     * IMPORTANT: This is not a normal page, it's an ACTION. This is where the "delete a song" button on songs/index
-     * directs the user after the click. This method handles all the data from the GET request (in the URL!) and then
-     * redirects the user back to songs/index via the last line: header(...)
-     * This is an example of how to handle a GET request.
-     * @param int $song_id Id of the to-delete song
-     */
-    public function delete($id)
-    {
-        // if we have an id of a song that should be deleted
-        if (isset($id)) {
-            // Instance new Model (Song)
-            $Song = new Song();
-            // do deleteSong() in model/model.php
-            $Song->deleteSong($id);
-        }
-
-        // where to go after song has been deleted
-        Redirect::to('songs/index');
-    }
-}
