+++ application/Core/Application.php
@@ -25,7 +25,8 @@ class Application
         // check for controller: no controller given ? then load start-page
         if (!$this->url_controller) {
 
-            $page = new \Mini\Controller\HomeController();
+            // NOTE: Sets default controller to Requests
+            $page = new \Mini\Controller\RequestsController();
             $page->index();
 
         } elseif (file_exists(APP . 'Controller/' . ucfirst($this->url_controller) . 'Controller.php')) {
