+++ application/view/layouts/sidebar.php
@@ -1,0 +1,39 @@
+<h2 class="sidenav__heading">FLEET</h2>
+<div class="sidenav__content">
+    <nav class="nav">
+        <ul>
+            <h3 class="sidenav__group--head">Requests: </i></h3>
+            <li>
+                <a href="<?= URL ?>requests" class="navlink active">Request Management</a>
+            </li>
+            <li>
+                <a href="#" class="navlink">Request History</a>
+            </li>
+            <h3 class="sidenav__group--head">Vehicles:</i>
+            </h3>
+            <li>
+                <a href="#" class="navlink">Vehicle Management</a>
+            </li>
+            <li>
+                <a href="#" class="navlink">Maintenance History</a>
+            </li>
+            <h3 class="sidenav__group--head">Drivers:</h3>
+            <li>
+                <a href="#" class="navlink">Driver Management</a>
+            </li>
+            <h3 class="sidenav__group--head">Users:</h3>
+            <li>
+                <a href="#" class="navlink">User Management</a>
+            </li>
+            <h3 class="sidenav__group--head">Journeys:</h3>
+            <li>
+                <a href="#" class="navlink">Journey Logs</a>
+            </li>
+            <li>
+                <a href="<?= URL ?>login/logout" class="navlink logout">Logout
+                    <i class="ion-android-exit icon-small"></i>
+                </a>
+            </li>
+        </ul>
+    </nav>
+</div>
